Version 1.1.1
Version 1.1.1
Changed energy usage from int to long to fix unable to store value error
Added Account Parameter: electricityCost
Added Channel energyCost
Added Channel powerCost
Added Channel energyReset
Added Channel voltageReset (use with caution, only when the lightwave value is messed up)
Added error logging for Message Failed return from the LightwaveRf servers

+++ src/main/java/org/openhab/binding/lightwaverf/internal/api/FeatureStatus.java
@@ -29,7 +29,7 @@ public class FeatureStatus {
 private String featureId;
 @SerializedName("value")
 @Expose
-private Integer value;
+private long value;
 
 public String getFeatureId() {
 return featureId;
@@ -39,11 +39,11 @@ public void setFeatureId(String featureId) {
 this.featureId = featureId;
 }
 
-public Integer getValue() {
+public long getValue() {
 return value;
 }
 
-public void setValue(Integer value) {
+public void setValue(long value) {
 this.value = value;
 }
 }
